requirements.txt
python-telegram-bot==21.4
llama-cpp-python==0.2.84